PVS-V1028 Possible overflow. Casting operands, not the result

+++ src/preview.cpp
@@ -1012,7 +1012,7 @@ namespace {
 
     DataBuf makePnm(uint32_t width, uint32_t height, const DataBuf &rgb)
     {
-        const long expectedSize = static_cast<long>(width * height * 3);
+        const long expectedSize = static_cast<long>(width) * static_cast<long>(height) * 3;
         if (rgb.size_ != expectedSize) {
 #ifndef SUPPRESS_WARNINGS
             EXV_WARNING << "Invalid size of preview data. Expected " << expectedSize << " bytes, got " << rgb.size_ << " bytes.\n";
